+++ app/texx/texxfuncs.go
